backend/requirements.txt
Flask==3.1.0
Pillow==11.1.0
torch==2.4.1
torchvision==0.19.1
Werkzeug==3.1.3
gunicorn==19.9.0
